[refactor] pokemon_list 패키지 명 list 로 변경
