+++ feature-jungwon/src/main/kotlin/com/samchi/feature/jungwon/presentation/JungwonRoute.kt
@@ -3,7 +3,7 @@ package com.samchi.feature.jungwon.presentation
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import com.samchi.feature.jungwon.presentation.pokemon_list.PokemonListScreen
+import com.samchi.feature.jungwon.presentation.list.PokemonListScreen
 
 @Composable
 fun JungwonRoute(
